* extract c3p0-specific part of DBPlugin to a separate class C3P0Plugin * c3p0 is still available if anybody should still need it
